+++ tests/test_web/test_webapi.py
@@ -6,18 +6,18 @@ from _pytest import monkeypatch
 import os
 
 import tidy3d as td
-import tidy3d.web as web
 
 from responses import matchers
 
+from tidy3d import Simulation
 from tidy3d.exceptions import SetupError
 from tidy3d.web.environment import Env
 from tidy3d.web.webapi import delete, delete_old, download, download_json, run
 from tidy3d.web.webapi import download_log, estimate_cost, get_info, get_run_info, get_tasks
 from tidy3d.web.webapi import load, load_simulation, start, upload, monitor, real_cost
 from tidy3d.web.container import Job, Batch
-from tidy3d.web.task import TaskInfo
 from tidy3d.web.asynchronous import run_async
+from tidy3d.web.file_util import compress_file_to_gzip
 
 from tidy3d.__main__ import main
 
@@ -25,6 +25,8 @@ from ..utils import TMP_DIR
 
 # variables used below
 FNAME_TMP = os.path.join(TMP_DIR, "web_test_tmp.json")
+HDF5_TMP = os.path.join(TMP_DIR, "web_test_tmp.hdf5")
+HDF5_GZ_TMP = os.path.join(TMP_DIR, "simulation.hdf5.gz")
 
 TASK_NAME = "task_name_test"
 TASK_ID = "1234"
@@ -80,6 +82,7 @@ def mock_upload(monkeypatch, set_api_key):
                     "callbackUrl": None,
                     "simulationType": "tidy3d",
                     "parentTasks": None,
+                    "fileType": "Gz",
                 }
             )
         ],
@@ -96,7 +99,7 @@ def mock_upload(monkeypatch, set_api_key):
     def mock_download(*args, **kwargs):
         pass
 
-    monkeypatch.setattr("tidy3d.web.simulation_task.upload_string", mock_download)
+    monkeypatch.setattr("tidy3d.web.simulation_task.upload_file", mock_download)
 
 
 @pytest.fixture
@@ -151,12 +154,10 @@ def mock_start(monkeypatch, set_api_key, mock_get_info):
 
 @pytest.fixture
 def mock_monitor(monkeypatch):
-
     status_count = [0]
     statuses = ("upload", "running", "running", "running", "running", "running", "success")
 
     def mock_get_status(task_id):
-
         current_count = min(status_count[0], len(statuses) - 1)
         current_status = statuses[current_count]
         status_count[0] += 1
@@ -288,7 +289,6 @@ def _test_load(mock_load, mock_get_info):
 
 @responses.activate
 def test_delete(set_api_key, mock_get_info):
-
     responses.add(
         responses.DELETE,
         f"{Env.current.web_api_endpoint}/tidy3d/tasks/{TASK_ID}",
@@ -311,16 +311,18 @@ def test_estimate_cost(set_api_key, mock_get_info, mock_metadata):
 
 @responses.activate
 def test_download_json(monkeypatch, mock_get_info):
+    sim = make_sim()
+
     def mock_download(*args, **kwargs):
-        file_path = kwargs["to_file"]
-        with open(file_path, "w") as f:
-            f.write("0.3,5.7")
+        file_path = "simulation.hdf5"
+        sim.to_file(file_path)
+        compress_file_to_gzip(file_path, "simulation.hdf5.gz")
 
     monkeypatch.setattr("tidy3d.web.simulation_task.download_file", mock_download)
 
     download_json(TASK_ID, FNAME_TMP)
     with open(FNAME_TMP, "r") as f:
-        assert f.read() == "0.3,5.7"
+        assert Simulation.parse_raw(f.read()) == sim
 
 
 @responses.activate
@@ -382,7 +384,6 @@ def test_delete_old(set_api_key):
 
 @responses.activate
 def test_get_tasks(set_api_key):
-
     responses.add(
         responses.GET,
         f"{Env.current.web_api_endpoint}/tidy3d/project",
@@ -423,7 +424,6 @@ def test_real_cost(mock_get_info):
 
 @responses.activate
 def test_job(mock_webapi, monkeypatch):
-
     monkeypatch.setattr("tidy3d.web.container.Job.load", lambda *args, **kwargs: True)
     sim = make_sim()
     j = Job(simulation=sim, task_name=TASK_NAME, folder_name=PROJECT_NAME)
@@ -444,7 +444,6 @@ def mock_job_status(monkeypatch):
 
 @responses.activate
 def test_batch(mock_webapi, mock_job_status):
-
     # monkeypatch.setattr("tidy3d.web.container.Batch.monitor", lambda self: time.sleep(0.1))
     # monkeypatch.setattr("tidy3d.web.container.Job.status", property(lambda self: "success"))
 
@@ -460,7 +459,6 @@ def test_batch(mock_webapi, mock_job_status):
 
 @responses.activate
 def test_async(mock_webapi, mock_job_status):
-
     # monkeypatch.setattr("tidy3d.web.container.Job.status", property(lambda self: "success"))
     sims = {TASK_NAME: make_sim()}
     batch_data = run_async(sims, folder_name=PROJECT_NAME)
@@ -471,7 +469,6 @@ def test_async(mock_webapi, mock_job_status):
 
 @responses.activate
 def test_main(mock_webapi, monkeypatch, mock_job_status):
-
     # sims = {TASK_NAME: make_sim()}
     # batch_data = run_async(sims, folder_name=PROJECT_NAME)